forced call of basic post_process_lane_export signature
background: In the automatic tracking flavours we use the plugins to export.
Those have this special feature that that seem to car about file overwrites.
In manual tracking this is not implemented.
As DataExportGui checks for postprocessCanCheckForExistingFiles, and if true calls
the extended signature of post_process_lane_export with the checkOverWriteFiles
argument, postProcessCanCheckForExistingFiles is set to False here, in order to
force the basic signature, which is implemented in manualTrackingWorkflow