refactor: use elf instead of nifty directly in multicut
also deprecates the "old" default `Nifty_FmGreedy` solver in favor of the
new default `kerningham-lin`. Old projects will still be able to use the
old solver, and it's also still available in debug mode.

In debug mode also all multicut solvers available from elf are available.

Also removed old_div, as it's not doing anything to numpy arrays and
was not necessary in the first place.